+++ src/test/java/aoc2024/Day6Test.java
@@ -24,22 +24,22 @@ class Day6Test extends DayTest {
     }
 
     @Override
-    protected int part1ExampleResult() {
+    protected long part1ExampleResult() {
         return 41;
     }
 
     @Override
-    protected int part1Result() {
+    protected long part1Result() {
         return 5131;
     }
 
     @Override
-    protected int part2ExampleResult() {
+    protected long part2ExampleResult() {
         return 6;
     }
 
     @Override
-    protected int part2Result() {
+    protected long part2Result() {
         return 1784;
     }
 }
